fix: restore ValidatingForm, add validation SCSS
