+++ apps/site/src/lib/components/ValidatingForm/ValidatingForm.module.scss
@@ -1,0 +1,9 @@
+.validated {
+	&:invalid p {
+		display: block;
+	}
+}
+
+.notYetValidated p {
+	display: none;
+}
